group project completed semi-finally
