+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.academy.techcenture</groupId>
-    <artifactId>e2e-using-page-object-model</artifactId>
+    <artifactId>e2e-using-page-object-model-extent-reports</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
