+++ src/main/java/com/academy/techcenture/ecommerce/utils/CommonUtils.java
@@ -1,7 +1,14 @@
 package com.academy.techcenture.ecommerce.utils;
 
 import com.github.javafaker.Faker;
+import org.apache.pdfbox.pdmodel.PDDocument;
+import org.apache.pdfbox.text.PDFTextStripper;
+
+import java.io.File;
+import java.io.IOException;
 import java.time.LocalDate;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.Locale;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
@@ -95,6 +102,37 @@ public class CommonUtils {
         return (int)(Math.random() * (5 - 1  + 1 ) + 1);
     }
 
+    public String getRecentFileName(String directoryPath) {
+
+        File file = new File(directoryPath);
+
+        String[] directories = file.list((current, name) -> new File(current, name).isDirectory());
+
+        Arrays.sort(directories, Collections.reverseOrder());
+
+        String latestFileName = directories[0];
+
+        File loc = new File(directoryPath + "/" + latestFileName) ;
+
+        File[] files = loc.listFiles();
+
+        String fileName = files[0].getAbsolutePath();
+
+        System.out.println(fileName);
+
+        return fileName;
+    }
+
+    public String readPdfDocument(String pdfFilePath) throws IOException {
+        String content  = "";
+        PDDocument document = PDDocument.load(new File(pdfFilePath));
+        if (!document.isEncrypted()) {
+            PDFTextStripper stripper = new PDFTextStripper();
+            content = stripper.getText(document);
+        }
+        document.close();
+        return content;
+    }
 
 
 }
