+++ src/test/java/com/academy/techtenture/ecommerce/base/BaseTest.java
@@ -82,7 +82,6 @@ public abstract class BaseTest {
         }
     }
 
-
     public static String getScreenshot(WebDriver driver, String screenshotName) throws IOException {
         String dateName = new SimpleDateFormat("yyyyMMddhhmmss").format(new Date());
         TakesScreenshot ts = (TakesScreenshot) driver;
